Creating option via configuration to refresh cache or invalidate when there are product modifications

+++ src/CacheInvalidation.Api/UseCases/Handlers/ProductUpdatedEventHandler.cs
@@ -11,24 +11,32 @@ namespace CacheInvalidation.Api.UseCases.Handlers
         private readonly ICacheDatabase _cacheDatabase;
         private readonly CacheConfig _cacheConfig;
         private readonly ILogger<ProductUpdatedEventHandler> _logger;
+        private readonly RefreshProductCache _refreshProductCache;
 
-        public ProductUpdatedEventHandler(ICacheDatabase cacheDatabase, IConfiguration configuration, ILogger<ProductUpdatedEventHandler> logger)
+        public ProductUpdatedEventHandler(ICacheDatabase cacheDatabase, IConfiguration configuration, ILogger<ProductUpdatedEventHandler> logger, RefreshProductCache refreshProductCache)
         {
             this._cacheDatabase = cacheDatabase;
             this._cacheConfig = configuration.GetSection("CacheConfig").Get<CacheConfig>();
             this._logger = logger;
+            this._refreshProductCache = refreshProductCache;
         }
 
-        public Task HandleAsync(ProductUpdatedEvent notification, CancellationToken cancellationToken = default)
+        public async Task HandleAsync(ProductUpdatedEvent notification, CancellationToken cancellationToken = default)
         {
-            this._cacheDatabase.RemoverAsync(this._cacheConfig.ProductCacheKey, cancellationToken);
+            if (this._cacheConfig.ItsToRefresh.GetValueOrDefault() == true)
+            {
+                await this._refreshProductCache.ExecuteAsync(cancellationToken);
+            }
+            else
+            {
+                await this._cacheDatabase.RemoverAsync(this._cacheConfig.ProductCacheKey, cancellationToken);
+            }
             var output = new OutboxMessage(notification.Product.Id,
                 nameof(ProductUpdatedEvent).ToString(),
                 JsonSerializer.Serialize(notification),
                 DateTime.UtcNow,
                 false);
             this._logger.LogInformation($"Product updated event handled: {JsonSerializer.Serialize(output)}");
-            return Task.CompletedTask;
         }
     }
 }
